댓글 CRUD 공통 처리 예외

+++ src/main/java/com/example/scheduleapp/comment/service/CommentServiceImpl.java
@@ -2,11 +2,13 @@ package com.example.scheduleapp.comment.service;
 
 import com.example.scheduleapp.comment.dto.response.CommentResponseDto;
 import com.example.scheduleapp.comment.entity.Comment;
+import com.example.scheduleapp.global.exception.ErrorCode;
+import com.example.scheduleapp.global.exception.custom.EntityNotFoundException;
+import com.example.scheduleapp.global.exception.custom.ForbiddenException;
 import com.example.scheduleapp.member.entity.Member;
 import com.example.scheduleapp.member.service.MemberServiceImpl;
 import com.example.scheduleapp.schedule.entity.Schedule;
 import com.example.scheduleapp.comment.repository.CommentRepository;
-import com.example.scheduleapp.schedule.repository.ScheduleRepository;
 import com.example.scheduleapp.schedule.service.ScheduleServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,22 +20,22 @@ import org.springframework.web.server.ResponseStatusException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class CommentServiceImpl implements CommentService {
+public class CommentServiceImpl {
 
-    //Service가 있어야 한다.
     private final MemberServiceImpl memberService;
     private final ScheduleServiceImpl scheduleService;
     private final CommentRepository commentRepository;
 
-    @Override
     @Transactional
     public CommentResponseDto createComment(Long scheduleId, Long httpSessionId, String content) {
-        //유저와 일정이 모두 있어야 댓글이 존재할 수 있다. (또한 유저는 일정을 선택하여 댓글 등록을 할 수 있다.)
-        Member member = memberService.getUserByIdOrElseThrow(httpSessionId);
+
+        Member member = memberService.getUserById(httpSessionId);
         Schedule schedule = scheduleService.findScheduleById(scheduleId);
 
         log.info("유저 및 일정 확인 완료");
@@ -44,6 +46,7 @@ public class CommentServiceImpl implements CommentService {
         //값 저장하기
         Comment savedComment = commentRepository.save(comment);
 
+        /*todo: CommentResponseDto 중복 되는 부분 빼기*/
         log.info("댓글 생성 완료");
         return new CommentResponseDto(
                 savedComment.getId(),
@@ -55,20 +58,6 @@ public class CommentServiceImpl implements CommentService {
         );
     }
 
-    //todo: 댓글의 날짜 출력 형식 변경하기
-    @Override
-    @Transactional
-    public void updateComment(Long commentId, String comment) {
-        Comment findComment = commentRepository.findByIdOrElseThrow(commentId);
-
-        //todo: 입력한 댓글이 비어있다면 기존 값 유지
-        //댓글 수정
-        findComment.updateComment(comment);
-
-        log.info("댓글 수정 완료");
-    }
-
-    @Override
     public List<CommentResponseDto> getComments() {
         List<CommentResponseDto> commentList = commentRepository.findAll()
                 .stream()
@@ -76,21 +65,75 @@ public class CommentServiceImpl implements CommentService {
                 .toList();
 
         if (commentList.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "댓글이 존재하지 않습니다.");
+            throw new EntityNotFoundException(ErrorCode.NOT_FOUND);
         }
 
-        log.info("댓글 조회 완료");
+        log.info("댓글 전체 조회 완료");
 
         return commentList;
     }
 
-    @Override
-    public void deleteComment(Long commentId) {
-        Comment findComment = commentRepository.findByIdOrElseThrow(commentId);
+    public CommentResponseDto getComment(Long commentId) {
+        Comment commentById = findCommentById(commentId);
+
+        log.info("댓글 단건 조회 완료");
+
+        return new CommentResponseDto(
+                commentById.getId(),
+                commentById.getMember().getId(),
+                commentById.getSchedule().getId(),
+                commentById.getContent(),
+                localDateTimeFormat(commentById.getCreatedAt()),
+                localDateTimeFormat(commentById.getUpdatedAt())
+        );
+
+    }
+
+    //todo: 댓글의 날짜 출력 형식 변경하기
+    @Transactional
+    public CommentResponseDto updateComment(Long httpSessionId, Long commentId, String comment) {
+        Comment findComment = findCommentById(commentId);
+
+        /*권한이 없는 경우*/
+        if (!Objects.equals(httpSessionId, findComment.getMember().getId())) {
+            throw new ForbiddenException(ErrorCode.CANNOT_UPDATE_OTHERS_DATA);
+        }
+
+        //todo: 입력한 댓글이 비어있다면 기존 값 유지
+        if (!findComment.getContent().equals(comment)) {
+            findComment.updateComment(comment);
+        }
+
+        log.info("댓글 수정 완료");
+
+        return new CommentResponseDto(
+                findComment.getId(),
+                findComment.getMember().getId(),
+                findComment.getSchedule().getId(),
+                findComment.getContent(),
+                localDateTimeFormat(findComment.getCreatedAt()),
+                localDateTimeFormat(findComment.getUpdatedAt())
+        );
+    }
+
+    public void deleteComment(Long httpSessionId,Long commentId) {
+        Comment findComment = findCommentById(commentId);
+
+        if(!Objects.equals(httpSessionId, findComment.getMember().getId())) {
+            throw new ForbiddenException(ErrorCode.CANNOT_UPDATE_OTHERS_DATA);
+        }
 
         //댓글 삭제
         commentRepository.delete(findComment);
         log.info("댓글 삭제 완료");
+    }
+
+    private Comment findCommentById(Long commentId) {
+        Optional<Comment> commentbyId = commentRepository.findById(commentId);
+        if (commentbyId.isEmpty()) {
+            throw new EntityNotFoundException(ErrorCode.NOT_FOUND);
+        }
+        return commentbyId.get();
     }
 
     private String localDateTimeFormat(LocalDateTime dateTime) {
